rename api functions and remove use of _config_ noun
