+++ pkg/services/publicdashboards/models/models.go
@@ -151,7 +151,7 @@ func (pd PublicDashboard) BuildTimeSettings(dashboard *models.Dashboard) TimeSet
 }
 
 // DTO for transforming user input in the api
-type SavePublicDashboardConfigDTO struct {
+type SavePublicDashboardDTO struct {
 	DashboardUid    string
 	OrgId           int64
 	UserId          int64
@@ -172,6 +172,6 @@ type AnnotationsQueryDTO struct {
 // COMMANDS
 //
 
-type SavePublicDashboardConfigCommand struct {
+type SavePublicDashboardCommand struct {
 	PublicDashboard PublicDashboard
 }
